Port render_debug cylinder drawer FUN_001896d0 (type 4)
Draws or caches a debug cylinder. With flag clear, caches a type-4 primitive
(base, height vector, radius, color). With flag set, calls the cylinder builder
FUN_00188d00 to fill two 17-vertex circle rings, then draws the wireframe:
  - each ring as a 16-segment line strip;
  - four vertical edges connecting the two rings (every 4th vertex);
  - two diameter cross-lines per ring (vertex i to vertex i+8).
The two rings share one contiguous 102-float buffer (buf1 = [0..50],
buf2 = [51..101]); MSVC places it at the original's exact stack offsets
(-0x198 / -0xcc), so the cylinder-builder call site matches byte-for-byte.

Loop counts, strides and buffer offsets verified against the disasm (0x1896d0).
center/height_vec/color are void* (matching the placeholder its callers -- the
cache flush and collision_usage -- were written against), so only the radius
argument, which both passed as int-bits (*(int*)...), was changed to the float
field / *(float*) at those two call sites.

Evidence:
- Build PASS and links cleanly (activating this 5-arg cdecl drawer does not hit
  the cylinder builder's activation link break; it calls the dormant builder via
  its @reg thunk).
- VC71 68.8% (125/131 insns). --show-diffs confirms the structure matches: the
  cylinder call sets ECX/EDX to the ring buffers at the original offsets and the
  line loops address the correct buffer vertices; the gap is the C using indexed
  buffer access where the original walks a byte pointer, plus register/schedule.
- Runtime not independently checkable here (the builder it calls is dormant, so
  the equivalence harness stubs it); rests on the disasm trace + structural
  match.
- ABI audit PASS, hazard scan clean.

+++ src/halo/physics/collision_usage.c
@@ -89,7 +89,7 @@ void FUN_0014c7b0(int16_t *param_1)
       iVar1 = (int)sVar2;
       FUN_001896d0(
         1, param_1 + iVar1 * 0x14 + 0xe0a, param_1 + iVar1 * 0x14 + 0xe10,
-        *(int *)(param_1 + iVar1 * 0x14 + 0xe16), *(void **)0x2ee6d4);
+        *(float *)(param_1 + iVar1 * 0x14 + 0xe16), *(void **)0x2ee6d4);
       sVar2 = sVar2 + 1;
     } while (sVar2 < param_1[1]);
   }
